Create Sync-Folders.ps1
I've been overwhelmed with how fast copy-item on powershell is! I was actually trying to figure out somehow if I can figure out a useable 'copy-rawitem' to use as little overhead as possible but with the results I got from copy-item (I used measure-command everytime I copied) there was no point. It was already very satisfactory for what I needed.

I used this to do the initial 'sync' from my desktop to my nas-box (dns-323) then using synctoy on a schedule to maintain sync on them so I don't have to worry about data missing on one or the other. I particularly liked the 'touch' function.I was dealing mostly with media files so I am fairly confident that a 2gig video file is exactly the same as it's counterpart on the other machine as long as they have the same size and name and overwriting that is costly, especially for my poor 64MB dns-323. I think I was able to make synctoy 'skip' at least half of what it was supposed to 'overwrite' just cause they had different mod times.